fix: Remove extraneous "KLM" characters from beginning of NIE check
If such characters were present in the NIE to be checked a TypeError
would be thrown:

TypeError: Unsupported operand types: string % int

Tried to find any evidence of these charatcters actually being allowed
for Spanish NIE but no evidence of such a thing was found. If this were
an error and they are in fact supported then proper handling of those
inputs is advised.